// src/app/main/logistics/material-request-management/shared/material-proposed-table-row.interface.ts
export interface MaterialProposedTableRow {
  id?: number;
  pn?: string;
  designation?: string;
  sn?: string;
  site?: string;
  warehouse?: string;
  location?: string;
  responsible?: string;
  potential?: string;
  expiration?: string;
  satisfactionRate?: number;
  status?: string;
  message?: string;
  satisfactionColor?: string;
  interchangebility?: string;
  buttonName?: string;
}